Move content validation to shared markdown package
